requirements.txt
ultralytics
opencv-python
python-dotenv
torch
timm
scikit-learn
scipy
